requirements/base.txt
django
python-dotenv
djangorestframework
django-autoslug
django-countries
django-phonenumber-field
phonenumbers
drf-yasg
argon2-cffi
djoser
pillow
django-filter
django-celery-email
django-celery-beat
redis
celery
flower
django-taggit
boto3
django-storages